requirements.txt
pprint
flask
pymongo[srv]
Flask-Pymongo
flask_cors
json
os
python-dotenv
bson